pyproject.toml
[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[project]
name = "summary"
version = "0.0.5"
description = "Summary application to summarize the scraped data"
authors = [
    { name = "Roman", email = "iamromandev@gmail.com" },
]
readme = "README.md"
requires-python = "==3.12.10"
dependencies = [
    "loguru==0.7.3",
    "fastapi[all]==0.115.12",
    "tortoise-orm[accel,aiomysql]==0.25.1",
    "aerich[toml,mysql]==0.9.1",
    "redis==6.2.0",
    "httpx==0.28.1",
    "spacy==3.8.6",
    "toml==0.10.2",
    "beautifulsoup4==4.13.4",
    "aiohttp==3.12.13",
    "trafilatura==2.0.0",
]

[dependency-groups]
dev = [
    "pre-commit==4.2.0",
    "ruff==0.11.11",
    "mypy==1.15.0",
    "ty==0.0.1a6",
]

[tool.uv]
package = false

[tool.aerich]
tortoise_orm = "src.db.DB_CONFIG"
location = "./src/db/migrations"
src_folder = "./."
